requirements.txt
fastapi==0.99.1
psycopg2-binary==2.9.6
pytest==7.4.0
SQLAlchemy==2.0.17
uvicorn==0.22.0
alembic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
jose-python